[OCO-21] Marketplace - Adiciona Tela de Produtos

+++ app-marketplace/marketplace-backend/Backend.Domain/Models/SubCategory.cs
@@ -5,8 +5,8 @@ namespace Backend.Domain.Models
 {
     public class SubCategory
     {
-        public int Id { get; set; }
-        public int CategoryId { get; set; }
+        public long Id { get; set; }
+        public long CategoryId { get; set; }
         public string? Name { get; set; }
 
         public virtual Category? Category { get; set; }
